StochasticGroundMotionSimulation

Documentation for the Julia package StochasticGroundMotionSimulation.jl. The main module StochasticGroundMotionSimulation provides an interface to the stochastic method for the simulation of response spectral ordinates via Random Vibration Theory.

Example

A number of default parameters are already set within the package.

using StochasticGroundMotionSimulation

Δσ = 100.0
Rrefi = [ 1.0, 50.0, Inf ]
γi = [ 1.0, 0.5 ]
Q0 = 200.0
η = 0.5
κ0 = 0.039

src = SourceParameters(Δσ)
geo = GeometricSpreadingParameters(Rrefi, γi, :CY14)
sat = NearSourceSaturationParameters(:BT15)
ane = AnelasticAttenuationParameters(Q0, η)
path = PathParameters(geo, sat, ane)
site = SiteParameters(κ0, :Boore2016)

fas = FourierParameters(src, path, site)

rvt = RandomVibrationParameters()

T = 1.0
m = 6.0
r_rup = 10.0
r_ps = equivalent_point_source_distance(r_rup, m, fas)

Sa = rvt_response_spectral_ordinate(T, m, r_ps, fas, rvt)
println("Sa = $(Sa)g, for m = $m, r_rup = $r_rup, and a period of T = $(T) s")

Sa = 0.11691325947492309g, for m = 6.0, r_rup = 10.0, and a period of T = 1.0 s

Fourier Parameters

Source Parameters

Path Parameters

Site Parameters

Fourier Amplitude Spectrum

StochasticGroundMotionSimulation

Example

Path Scaling

Geometric Spreading

Near Source Saturation

Anelastic Attenuation

Random Vibration Theory Parameters

Single Degree of Freedom Oscillator Parameters

Site Scaling

Fourier Source Spectrum

StochasticGroundMotionSimulation.AnelasticAttenuationParameters

StochasticGroundMotionSimulation.FourierParameters

StochasticGroundMotionSimulation.GeometricSpreadingParameters

StochasticGroundMotionSimulation.NearSourceSaturationParameters

StochasticGroundMotionSimulation.Oscillator

StochasticGroundMotionSimulation.Oscillator

StochasticGroundMotionSimulation.PathParameters

StochasticGroundMotionSimulation.RandomVibrationParameters

StochasticGroundMotionSimulation.SiteParameters

StochasticGroundMotionSimulation.SourceParameters

StochasticGroundMotionSimulation.alatik_2021_cy14_inverted_amplification

StochasticGroundMotionSimulation.alatik_2021_cy14_inverted_amplification_seg

StochasticGroundMotionSimulation.anelastic_attenuation

StochasticGroundMotionSimulation.boore_2016_generic_amplification

StochasticGroundMotionSimulation.combined_kappa_frequency

StochasticGroundMotionSimulation.corner_frequency

StochasticGroundMotionSimulation.corner_frequency_atkinson_silva_2000

StochasticGroundMotionSimulation.corner_frequency_brune

StochasticGroundMotionSimulation.equivalent_point_source_distance

StochasticGroundMotionSimulation.finite_fault_factor_bt15

StochasticGroundMotionSimulation.finite_fault_factor_cy14

StochasticGroundMotionSimulation.finite_fault_factor_ya15

StochasticGroundMotionSimulation.fourier_attenuation

StochasticGroundMotionSimulation.fourier_constant

StochasticGroundMotionSimulation.fourier_path

StochasticGroundMotionSimulation.fourier_site

StochasticGroundMotionSimulation.fourier_source

StochasticGroundMotionSimulation.fourier_source

StochasticGroundMotionSimulation.fourier_source_shape

StochasticGroundMotionSimulation.fourier_source_shape

StochasticGroundMotionSimulation.fourier_source_shape

StochasticGroundMotionSimulation.fourier_spectral_ordinate

StochasticGroundMotionSimulation.fourier_spectral_ordinate

StochasticGroundMotionSimulation.fourier_spectral_ordinate

StochasticGroundMotionSimulation.fourier_spectrum

StochasticGroundMotionSimulation.fourier_spectrum!

StochasticGroundMotionSimulation.geometric_spreading

StochasticGroundMotionSimulation.geometric_spreading

StochasticGroundMotionSimulation.geometric_spreading_cy14

StochasticGroundMotionSimulation.geometric_spreading_cy14mod

StochasticGroundMotionSimulation.geometric_spreading_piecewise

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.get_parametric_type

StochasticGroundMotionSimulation.kappa_filter

StochasticGroundMotionSimulation.magnitude_to_moment

StochasticGroundMotionSimulation.near_source_saturation

StochasticGroundMotionSimulation.near_source_saturation

StochasticGroundMotionSimulation.near_source_saturation

StochasticGroundMotionSimulation.peak_factor

StochasticGroundMotionSimulation.peak_factor

StochasticGroundMotionSimulation.period

StochasticGroundMotionSimulation.period

StochasticGroundMotionSimulation.rms_duration

StochasticGroundMotionSimulation.rupture_distance_from_equivalent_point_source_distance

StochasticGroundMotionSimulation.simpsons_rule

StochasticGroundMotionSimulation.site_amplification

StochasticGroundMotionSimulation.site_amplification

StochasticGroundMotionSimulation.site_amplification

StochasticGroundMotionSimulation.squared_fourier_spectral_ordinate

StochasticGroundMotionSimulation.squared_fourier_spectrum

StochasticGroundMotionSimulation.squared_fourier_spectrum!

StochasticGroundMotionSimulation.squared_transfer

StochasticGroundMotionSimulation.squared_transfer

StochasticGroundMotionSimulation.squared_transfer!

StochasticGroundMotionSimulation.squared_transfer!

StochasticGroundMotionSimulation.transfer

StochasticGroundMotionSimulation.transfer

StochasticGroundMotionSimulation.transfer

StochasticGroundMotionSimulation.transfer!

StochasticGroundMotionSimulation.trapezoidal_rule

StochasticGroundMotionSimulation.unit_generic_amplification

StochasticGroundMotionSimulation.FourierParameters — Type

FourierParameters

Custom type for the parameters the Fourier amplitude spectrum. This type is comprised of source, path and site types, and so has a base constructor of: FourierParameters(src::S, path::T, site::U) where {S<:SourceParameters, T<:PathParameters, U<:SiteParameters}

See also: SourceParameters, PathParameters, SiteParameters

StochasticGroundMotionSimulation.Oscillator — Type

Oscillator

Custom type to represent a SDOF oscillator. The type has two fields:

f_n is the natural frequency of the oscillator

ζ_n is the damping ratio

Examples

    sdof = Oscillator( 1.0, 0.05 )

StochasticGroundMotionSimulation.Oscillator — Method

Oscillator(f_n)

Default initializer setting the damping ratio to 5% of critical

f_n is the natural frequency of the oscillator (in Hz)

# Examples

    f_n = 2.0
    sdof = Oscillator(f_n)

StochasticGroundMotionSimulation.RandomVibrationParameters — Type

RandomVibrationParameters

Struct holding parameters/methods for Random Vibration Theory.

pf_method is the method used for peak factor computation

- `:DK80` (default) is Der Kiureghian (1980), building on Vanmarcke (1975)
- `:CL56` is Cartwright Longuet-Higgins (1956)

dur_ex is the model for excitation duration

- `:BT14` is the Boore & Thompson (2014) model - note that this is adpated to work with `r_ps`

dur_rms is the model for rms duration

- `:BT12` is the Boore & Thompson (2012) model

dur_region is the region specified for the duration model

- `:WNA` is western North America
- `:ENA` is eastern North America

StochasticGroundMotionSimulation.alatik_2021_cy14_inverted_amplification — Method

alatik_2021_cy14_inverted_amplification(f::T) where T<:Real

Computes the generic crustal amplification for a velocity profile with Vs30 of 760 m/s obtained from inverting the Chiou & Youngs (2014) ground-motion model.

Amplifications provided ahead of Al Atik & Abrahamson (2021) publication.

Examples

f = 5.0
	Af = alatik_2021_cy14_inverted_amplification(f)

StochasticGroundMotionSimulation.alatik_2021_cy14_inverted_amplification_seg — Method

alatik_2021_cy14_inverted_amplification_seg(f::T) where T<:Real

Computes the generic crustal amplification for a velocity profile with Vs30 of 760 m/s obtained from inverting the Chiou & Youngs (2014) ground-motion model.

Amplifications provided ahead of Al Atik & Abrahamson (2021) publication.

Same as alatik2021cy14invertedamplification, just segmented to speed up the indexing.

Examples

f = 5.0
	Af = alatik_2021_cy14_inverted_amplification_seg(f)

StochasticGroundMotionSimulation.anelastic_attenuation — Method

anelastic_attenuation(f::S, r::T, anelastic::AnelasticAttenuationParameters) where {S<:Float64,T<:Real}

Anelastic attenuation filter, computed using equivalent point source distance metric or a standard rupture distance.

StochasticGroundMotionSimulation.boore_2016_generic_amplification — Method

boore_2016_generic_amplification(f)

Computes the generic crustal amplification for a WUS velocity profile with Vs30 of 760 m/s using the Boore (2016) velocity model.

Examples

f = 5.0
	Af = boore_2016_generic_amplification(f)

StochasticGroundMotionSimulation.combined_kappa_frequency — Method

combined_kappa_frequency(r::T, Af2target::Float64, ane::AnelasticAttenuationParameters, site::SiteParameters) where T<:Real

Frequency at which the combined κr and κ0 filters (squared versions) give a value of Af2target. r can be either r_ps or r_rup depending upon what matches ane.rmetric

StochasticGroundMotionSimulation.corner_frequency — Method

corner_frequency(m::U, src::SourceParameters{S,T}) where {S<:Float64, T<:Real, U<:Real}

Computes a 3-tuple of corner frequency components, depending upon source spectrum type. By default the single-corner Brune spectrum is considered, but if src.model equals :Atkinson_Silva_2000 then the components of the double-corner spectrum are returned. If some other symbol is passed then the Brune model is returned.

Examples

    m = 6.0
    Δσ = 100.0
    κ0 = 0.035
    fas = FASParams(Δσ, κ0)
    # compute single corner frequency
	src.model = :Brune
    fc, tmp1, tmp2 = corner_frequency(m, src)
    # compute double corner frequencies
	src.model = :Atkinson_Silva_2000
    fa, fb, ε = corner_frequency(m, src)

StochasticGroundMotionSimulation.corner_frequency_atkinson_silva_2000 — Method

corner_frequency_atkinson_silva_2000(m::T) where T<:Real

Computes the corner frequencies, fa, fb, and the mixing parameter ε from the Atkinson & Silva (2000) double corner frequency model. This is the default source corner frequency model used by Boore & Thompson (2014) to define their source duration. But note that they just use fa. This function returns fb and ε also

# Examples

    m = 6.0
    fa, fb, ε = corner_frequency_atkinson_silva_2000(m)

See also: corner_frequency

StochasticGroundMotionSimulation.corner_frequency_brune — Method

corner_frequency_brune(m::S, Δσ::T, β::Float64=3.5) where {S<:Real, T<:Real}

Computes the corner frequency using the Brune model. - m is the moment magnitude - Δσ is the stress drop in units of bars - β is the shear-wave velocity at the source in units of km/s

# Examples

m = 6.0
	Δσ = 100.0
	β = 3.5
	fc = corner_frequency_brune(m, Δσ, β)

StochasticGroundMotionSimulation.equivalent_point_source_distance — Method

equivalent_point_source_distance(r, m, sat::NearSourceSaturationParameters)

Compute equivalent point source distance

r is $r_{hyp}$ or $r_{rup}$ (depending upon the size of the event – but is nominally $r_{rup}$)

m is magnitude

sat contains the NearSourceSaturationParameters

See also: near_source_saturation

StochasticGroundMotionSimulation.finite_fault_factor_bt15 — Method

finite_fault_factor_bt15(m::T) where T<:Real

Finite fault factor from Boore & Thompson (2015) used to create equivalent point-source distance.

See also: near_source_saturation, finite_fault_factor_ya15, finite_fault_factor_cy14

StochasticGroundMotionSimulation.finite_fault_factor_cy14 — Method

finite_fault_factor_cy14(m::T) where T<:Real

Finite fault factor for Chiou & Youngs (2014) used to create equivalent point-source distance. This is a model developed to match the average saturation length over the full period range.

See also: near_source_saturation, finite_fault_factor_bt15, finite_fault_factor_ya15

StochasticGroundMotionSimulation.finite_fault_factor_ya15 — Method

finite_fault_factor_ya15(m::T) where T<:Real

Finite fault factor from Yenier & Atkinson (2015) used to create equivalent point-source distance.

See also: near_source_saturation, finite_fault_factor_bt15, finite_fault_factor_cy14

StochasticGroundMotionSimulation.fourier_attenuation — Method

fourier_attenuation(f::S, r::T, ane::AnelasticAttenuationParameters{U,V}, site::SiteParameters{W}) where {S<:Float64,T<:Real,U<:Real,V<:Real,W<:Real}

Combined full-path attenuation, including Q(f) effects and κ0 filter for frequency f Distance defined in terms of an equivalent point source distance r_ps or rupture distance r_rup depending upon what metric is defined in ane.rmetric

StochasticGroundMotionSimulation.fourier_constant — Method

fourier_constant(src::SourceParameters)

Define the constant source term for the Fourier Amplitude Spectrum. Constant set to permit distances to be passed in km, densities in t/m^3, and velocities in km/s The reference distance is set to 1.0 km (and interpreted to be a rupture distance).

StochasticGroundMotionSimulation.fourier_path — Method

fourier_path(f::U, r_ps::T, m::S, geo::GeometricSpreadingParameters, sat::NearSourceSaturationParameters, ane::AnelasticAttenuationParameters) where {S<:Real,T<:Real,U<:Float64}

Path scaling of Fourier spectral model – combination of geometric spreading and anelastic attenuation.

See also: geometric_spreading, anelastic_attenuation

StochasticGroundMotionSimulation.fourier_site — Method

fourier_site(f::Float64, site::SiteParameters)

Combined site amplification and kappa filter for frequency f

StochasticGroundMotionSimulation.fourier_source — Method

fourier_source(f, m, fas::FourierParameters)

Source Fourier Amplitude Spectrum of displacement, without the constant term. This simply includes the seismic moment and the source spectral shape. Defined using a FourierParameters instance for the source model.

See also: fourier_source_shape

StochasticGroundMotionSimulation.fourier_source — Method

fourier_source(f::T, m::S, src::SourceParameters) where {S<:Real,T<:Float64}

Source Fourier Amplitude Spectrum of displacement, without the constant term. This simply includes the seismic moment and the source spectral shape.

See also: fourier_source_shape

StochasticGroundMotionSimulation.fourier_source_shape — Method

fourier_source_shape(f, m, fas::FourierParameters)

Source shape of the Fourier amplitude spectrum of displacement, without the constant term or seismic moment. Defined using a FourierParameters instance for the source model.

See also: fourier_source_shape

StochasticGroundMotionSimulation.fourier_source_shape — Method

fourier_source_shape(f::Float64, fa::T, fb::T, ε::T, model::Symbol) where T<:Real

Fourier amplitude spectral shape for displacement defined by corner frequencies.

See also: fourier_source_shape

StochasticGroundMotionSimulation.fourier_source_shape — Method

fourier_source_shape(f::T, m::S, src::SourceParameters) where {S<:Real,T<:Float64}

Source shape of the Fourier Amplitude Spectrum of displacement, without the constant term or seismic moment. This simply includes the source spectral shape.

StochasticGroundMotionSimulation.fourier_spectral_ordinate — Method

fourier_spectral_ordinate(f::U, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Float64}

Fourier acceleration spectral ordinate (m/s) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

StochasticGroundMotionSimulation.fourier_spectral_ordinate — Method

fourier_spectral_ordinate(f::S, m::S, r_ps::T, src::SourceParameters, geo::GeometricSpreadingParameters, ane::AnelasticAttenuationParameters, site::SiteParameters) where {S<:Float64,T<:Real}

Fourier acceleration spectral ordinate (m/s) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

src are the source parameters SourceParameters

geo are the geometric spreading parameters GeometricSpreadingParameters

sat are the near source saturation parameters NearSourceSaturationParameters

ane are the anelastic attenuation parameters AnelasticAttenuationParameters

site are the site parameters SiteParameters

See also: fourier_spectrum, fourier_spectrum!

StochasticGroundMotionSimulation.fourier_spectral_ordinate — Method

fourier_spectral_ordinate(f::U, m::S, r_ps::T, src::SourceParameters, path::PathParameters, site::SiteParameters) where {S<:Real,T<:Real,U<:Float64}

Fourier acceleration spectral ordinate (m/s) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

src are the source parameters SourceParameters

path are the path parameters PathParameters

site are the site parameters SiteParameters

StochasticGroundMotionSimulation.fourier_spectrum! — Method

fourier_spectrum!(Af::Vector{U}, f::Vector{V}, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Real,V<:Float64}

Fourier acceleration spectrum (m/s) based upon an equivalent point source distance r_ps

Af is the vector of fas amplitudes to be filled (m/s)

f is Vector of frequencies (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

See also: fourier_spectrum

StochasticGroundMotionSimulation.fourier_spectrum — Method

fourier_spectrum(f::Vector{U}, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Float64}

Fourier acceleration spectrum (m/s) based upon an equivalent point source distance r_ps

f is Vector of frequencies (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

See also: fourier_spectral_ordinate

StochasticGroundMotionSimulation.geometric_spreading — Method

geometric_spreading(r_ps::T, m::S, geo::GeometricSpreadingParameters, sat::NearSourceSaturationParameters) where {S<:Real, T<:Real}

Geometric spreading function, switches between different approaches on path.geo_model.

StochasticGroundMotionSimulation.geometric_spreading_cy14 — Method

geometric_spreading_cy14(r_ps::V, geo::GeometricSpreadingParameters{S,T,U}) where {S<:Float64, T<:Real, U<:AbstractVector{Bool}, V<:Real}

Geometric spreading function from Chiou & Youngs (2014). Defines a smooth transition from one rate γi[1] to another γi[2], with a spreading bandwidth of Rrefi[2] km.

StochasticGroundMotionSimulation.geometric_spreading_cy14mod — Method

geometric_spreading_cy14mod(r_ps::V, m::S, geo::GeometricSpreadingParameters{S,T,U}, sat::NearSourceSaturationParameters) where {S<:Float64, T<:Real, U<:AbstractVector{Bool}, V<:Real, W<:Real}

Geometric spreading function from Chiou & Youngs (2014). Modified to make use of both r_ps and r_rup so that only the first saturation term contaminates the source amplitudes. Defines a smooth transition from one rate γi[1] to another γi[2], with a spreading bandwidth of Rrefi[2] km.

StochasticGroundMotionSimulation.geometric_spreading_piecewise — Method

geometric_spreading_piecewise(r_ps::V, geo::GeometricSpreadingParameters{S,T,U}) where {S<:Float64, T<:Real, U<:AbstractVector{Bool}, V<:Real}

Piecewise linear (in log-log space) geometric spreading function. Makes use of the reference distances Rrefi and spreading rates γi in path.

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(fas::FourierParameters)

Extract type most elaborate type from parametric FourierParameters struct. This requires dropping down to lower level structs within fas.

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(path::PathParameters)

Extract type most elaborate type from parametric PathParameters struct. This requires dropping down to lower level structs within path.

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(anelastic::AnelasticAttenuationParameters{S,T}) where {S,T}

Extract type most elaborate type S or T from parametric AnelasticAttenuationParameters struct

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(geo::GeometricSpreadingParameters{S,T,U}) where {S,T,U} = T

Extract type of T from parametric GeometricSpreadingParameters struct

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(sat::NearSourceSaturationParameters{S,T,U}) where {S,T,U} = T

Extract type of T from parametric NearSourceSaturationParameters struct

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(site::SiteParameters{T}) where {T} = T

Extract type of T from parametric SiteParameters struct

StochasticGroundMotionSimulation.get_parametric_type — Method

get_parametric_type(src::SourceParameters{S,T}) where {S,T} = T

Extract type of T from parametric SourceParameters struct

StochasticGroundMotionSimulation.kappa_filter — Method

kappa_filter(f, site::SiteParameters)

Kappa filter for a given frequency f

StochasticGroundMotionSimulation.magnitude_to_moment — Method

magnitude_to_moment(m::T) where T<:Real

Converts moment magnitude to seismic moment (in dyne-cm).

# Examples

m = 6.0
	M0 = magnitude_to_moment(m)

StochasticGroundMotionSimulation.near_source_saturation — Method

near_source_saturation(m, fas::FourierParameters)

Near-source saturation term taking a FourierParameters struct.

See also: near_source_saturation

StochasticGroundMotionSimulation.near_source_saturation — Method

near_source_saturation(m, sat::NearSourceSaturationParameters)

Near-source saturation term. Used to create equivalent point-source distance. Switches methods based upon sat.model.

Arguments

Options for sat.model are:

:BT15 for Boore & Thompson (2015) finite fault factor

:YA15 for Yenier & Atkinson (2014) finite fault factor

:CY14 for a model fitted to the Chiou & Youngs (2014) saturation lengths (over all periods)

:None for zero saturation length

:ConstantConstrained for a constant value, sat.hconi[1], not subject to AD operations

:ConstantVariable for a constant value, sat.hvari[1], that is subject to AD operations

Any other symbol passed will return NaN.

See also: near_source_saturation

StochasticGroundMotionSimulation.near_source_saturation — Method

near_source_saturation(m, path::PathParameters)

Near-source saturation term taking a PathParameters struct.

See also: near_source_saturation

StochasticGroundMotionSimulation.peak_factor — Method

peak_factor(m::S, r_ps::T, fas::FourierParameters, sdof::Oscillator; pf_method::Symbol=:DK80) where {S<:Real,T<:Real}

Peak factor umax / urms with a switch of pf_method to determine the approach adopted. pf_method can currently be one of: - :CL56 for Cartright Longuet-Higgins (1956) - :DK80 for Der Kiureghian (1980), building on Vanmarcke (1975)

Defaults to :DK80.

StochasticGroundMotionSimulation.peak_factor — Method

peak_factor(m::S, r_ps::T, Dex::U, m0::V, fas::FourierParameters, sdof::Oscillator, rvt::RandomVibrationParameters) where {S<:Real,T<:Real,U<:Real,V<:Real}

Peak factor umax / urms with a switch of pf_method to determine the approach adopted. pf_method can currently be one of: - :CL56 for Cartright Longuet-Higgins (1956) - :DK80 for Der Kiureghian (1980), building on Vanmarcke (1975)

Defaults to :DK80.

StochasticGroundMotionSimulation.period — Method

period(sdof::Oscillator)

Natural period (s) of the sdof Oscillator.

StochasticGroundMotionSimulation.rms_duration — Method

rmsduration(m::S, rps::T, src::SourceParameters, sdof::Oscillator, rvt::RandomVibrationParameters) where {S<:Float64,T<:Real}

Returns a 3-tuple of (Drms, Dex, Dratio), using a switch on rvt.dur_rms. Default :BT12 makes use of the :BT14 model for excitation duration, Dex.

m is magnitude

r_ps is an equivalent point source distance

StochasticGroundMotionSimulation.rupture_distance_from_equivalent_point_source_distance — Method

rupture_distance_from_equivalent_point_source_distance(r_ps, m, sat::NearSourceSaturationParameters)

Compute rupture distance from equivalent point source distance

r_ps is the equivalent point source distance

m is magnitude

sat contains the NearSourceSaturationParameters

StochasticGroundMotionSimulation.simpsons_rule — Method

simpsons_rule(x::Vector, y::Vector)

Numerical integration via Simpson's rule.

Lengths of vectors must be equal to each other, and be odd (to have an even number of intervals). Values in x represent abcissa values, while y are the integrand counterparts.

StochasticGroundMotionSimulation.site_amplification — Method

site_amplification(f, fas::FourierParameters)

Computes the site amplification (impedance) for a given frequency f.

StochasticGroundMotionSimulation.site_amplification — Method

site_amplification(f, site::SiteParameters)

Computes the site amplification (impedance) for a given frequency f.

StochasticGroundMotionSimulation.site_amplification — Method

site_amplification(f::Real, amp_model::Symbol)

Computes the site amplification (impedance) for a given frequency f. Requires the keyword argument amp_model as a String and defaults to the Boore (2016) model. Currently, any other string passed to the function will return the unit amplification

# Examples

f = 5.0
	# returns the amplification from AlAtik (2021) in both cases
	Af = site_amplification(f)
    Af = site_amplification(f; amp_model=:AlAtik2021_cy14)
    # returns the Boore (2016) amplification
	Af = site_amplification(f; amp_model=:Boore2016)
	# returns 1.0
	Af = site_amplification(f; amp_model=:Unit)

StochasticGroundMotionSimulation.squared_fourier_spectral_ordinate — Method

squared_fourier_spectral_ordinate(f::U, m::S, r_ps::T, src::SourceParameters, geo::GeometricSpreadingParameters, ane::AnelasticAttenuationParameters, site::SiteParameters) where {S<:Real,T<:Real,U<:Float64}

Squared Fourier acceleration spectral ordinate (m^2/s^2) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

src are the source parameters SourceParameters

geo are the geometric spreading parameters GeometricSpreadingParameters

sat are the near source saturation parameters NearSourceSaturationParameters

ane are the anelastic attenuation parameters AnelasticAttenuationParameters

site are the site parameters SiteParameters

See also: fourier_spectral_ordinate, squared_fourier_spectrum

StochasticGroundMotionSimulation.squared_fourier_spectrum! — Method

squared_fourier_spectrum!(Afsq::Vector{U}, f::Vector{V}, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Real,V<:Float64}

Fourier acceleration spectrum (m/s) based upon an equivalent point source distance r_ps

Afsq is the vector of squared fas amplitudes to be filled (m^2/s^2)

f is Vector of frequencies (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

See also: squared_fourier_spectrum, 'squaredfourierspectral_ordinate'

StochasticGroundMotionSimulation.squared_fourier_spectrum — Method

squared_fourier_spectrum(f::Vector{U}, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Float64}

Squared Fourier amplitude spectrum. Useful within spectral moment calculations.

See also: fourier_spectrum, squared_fourier_spectral_ordinate

StochasticGroundMotionSimulation.squared_transfer! — Method

squared_transfer!(Hf2::Vector, f::Vector, sdof::Oscillator)

Computes the square of the modulus of the transfer function of a SDOF for a vector of frequencies in place: - Hf2::Vector is the pre-allocated vector into which the results are stored - f::Vector is the vector of frequencies - sdof::Oscillator is the oscillator instance

Examples

    f = collect(range(0.1, stop=10.0, step=0.01))
    sdof = Oscillator(1.0)
    Hf2 = similar(f)
    squared_transfer!(Hf2, f, sdof)

StochasticGroundMotionSimulation.squared_transfer! — Method

squared_transfer!(Hf2::Vector{T}, f::Vector{T}, sdof::Oscillator) where T<:Real

Computes the square of the modulus of the transfer function of a SDOF for a vector of frequencies in place: - Hf2::Vector is the pre-allocated vector into which the results are stored - f::Vector is the vector of frequencies - sdof::Oscillator is the oscillator instance Inputs derive from the Real type and so are differentiable.

Examples

    f = collect(range(0.1, stop=10.0, step=0.01))
    sdof = Oscillator(1.0)
    Hf2 = similar(f)
    squared_transfer!(Hf2, f, sdof)

StochasticGroundMotionSimulation.squared_transfer — Method

squared_transfer(f, sdof::Oscillator)

Compute the square of the transfer function for a SDOF system, sdof, at frequency f.

Examples

f = 2.0
	# create sdof with natural frequency f_n=1.0 and damping ζ=0.05
	sdof = Oscillator( 1.0, 0.05 )
	Hf2 = squared_transfer( f, sdof )

StochasticGroundMotionSimulation.transfer! — Method

transfer!(Hf::Vector{T}, f::Vector{T}, sdof::Oscillator) where T<:Real

Computes the modulus of the transfer function of a SDOF for a vector of frequencies in place - Hf::Vector is the pre-allocated vector into which the results are stored - f::Vector is the vector of frequencies - sdof::Oscillator is the oscillator instance

Examples

    f = collect(range(0.1, stop=10.0, step=0.01))
    sdof = Oscillator(1.0)
    Hf = similar(f)
    transfer!(Hf, f, sdof)

StochasticGroundMotionSimulation.transfer — Method

transfer(f::T, sdof::Oscillator) where {T<:Real}

Compute the modulus of the transfer function for a SDOF system

Examples

    f = 2.0
    sdof = Oscillator(1.0, 0.05)
    Hf = transfer(f, sdof)

StochasticGroundMotionSimulation.transfer — Method

transfer(f::Vector{T}, sdof::Oscillator) where T<:Real

Computes the modulus of the transfer function of a SDOF for a vector of frequencies

f::Vector is the vector of frequencies

sdof::Oscillator is the oscillator instance

Examples

  f = collect(range(0.1, stop=10.0, step=0.01))
  sdof = Oscillator(1.0)
  Hf = transfer(f, sdof)

StochasticGroundMotionSimulation.trapezoidal_rule — Method

trapezoidal_rule(x::Vector, y::Vector)

Numerical integration via the Trapezoidal rule.

Lengths of vectors must be equal to each other. Values in x represent equally-spaced abcissa values, while y are the integrand counterparts.

StochasticGroundMotionSimulation.unit_generic_amplification — Method

unit_generic_amplification()

Function used to represent null impedance effects. Use this function to represent site effects when no amplification is desired. Returns unity.

Examples

Af = unit_generic_amplification()
	# Af will equal 1.0