Fourier Parameters

Definition of the various custom types within the StochasticGroundMotionSimulation module. Types to store properties related to source, path, and site components of the Fourier spectral model are provided.

Source Parameters

The type SourceParameters holds the properties required to define the source spectrum of the Fourier Amplitude Spectrum.

StochasticGroundMotionSimulation.SourceParameters — Type

SourceParameters

Custom type defining the source parameters of a Fourier spectrum

Δσ::Union{Float64,Dual{Float64}} is the stress parameter in bars

RΘϕ::Float64 is the radiation pattern

V::Float64 is the partition factor (for splitting to horizontal components)

F::Float64 is the free surface factor

β::Float64 is the source velocity in units of km/s

ρ::Float64 is the source density in units of t/m^3 or g/cm^3

src_model::Symbol identifies the type of source spectrum (:Brune, AtkinsonSilva2000)

Path Parameters

The type PathParameters holds the properties required to define the path scaling of the Fourier Amplitude Spectrum.

StochasticGroundMotionSimulation.PathParameters — Type

PathParameters

Custom type defining the path parameters of a Fourier spectrum. Consists of three other custom structs

geometric is a GeometricSpreadingParameters type

saturation is a NearSourceSaturationParameters type

anelastic is an AnelasticAttenuationParameters type

This type also hold instances of three other custom types that define aspects of the path scaling:

GeometricSpreadingParameters defined in Geometric Spreading

NearSourceSaturationParameters defined in Near-source Saturation

AnelasticAttenuationParameters defined in Anelastic Attenuation

Geometric Spreading

StochasticGroundMotionSimulation.GeometricSpreadingParameters — Type

GeometricSpreadingParameters{S<:Real,T<:Real,U<:AbstractVector{Bool}}

Struct for geometric spreading parameters. Holds fields:

Rrefi are reference distances, these are <:Real but will generally be Float64 values

γconi are constant spreading rates, meaning that they will not be free for AD purposes

γvari are variable spreading rates, meaning that they can be represented as Dual numbers for AD

γfree is a vector of Bool instances, or a BitVector that indicates which segments are constant or variable. Variable spreading rates are given 1 or true

model is a symbol defining the type of spreading model :Piecewise, :CY14, :CY14mod

### Near-source Saturation

NearSourceSaturationParameters

Anelastic Attenuation

AnelasticAttenuationParameters

Site Parameters

The type SiteParameters holds information related to the site response – both impedance effects and damping.

SiteParameters