Fourier Amplitude Spectrum

The Fourier amplitude spectrum (FAS) can be represented as the product of source, path and site contributions.

Specifically, the Fourier amplitude spectrum $|A(f)|$ of acceleration (in units of m/s) is defined as:

\[|A(f; \bm{\theta})| = E(f; \bm{\theta}_E)\times P(f; \bm{\theta}_P) \times S(f; \bm{\theta}_S)\]

where $f$ is a frequency in Hz, and $\bm{\theta}$ holds all of the relevant model parameters and predictor variables. The Fourier Source Spectrum, $E(f; \bm{\theta}_E)$ is a function of the earthquake magnitude $m$, as well as other properties of the source. The Path Scaling, $P(f; \bm{\theta}_P)$ accounts for the effects of both geometric spreading and anelastic attenuation. The Site Scaling, $S(f; \bm{\theta}_S)$ includes the effects of near-surface impedance as well as damping ($\kappa_0$) effects.

Fourier spectral ordinates, or complete Fourier spectra are obtained via the functions:

StochasticGroundMotionSimulation.fourier_spectrum — Function

fourier_spectrum(f::Vector{U}, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Float64}

Fourier acceleration spectrum (m/s) based upon an equivalent point source distance r_ps

f is Vector of frequencies (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

See also: fourier_spectral_ordinate

StochasticGroundMotionSimulation.fourier_spectrum! — Function

fourier_spectrum!(Af::Vector{U}, f::Vector{V}, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Real,V<:Float64}

Fourier acceleration spectrum (m/s) based upon an equivalent point source distance r_ps

Af is the vector of fas amplitudes to be filled (m/s)

f is Vector of frequencies (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

See also: fourier_spectrum

StochasticGroundMotionSimulation.fourier_spectral_ordinate — Function

fourier_spectral_ordinate(f::S, m::S, r_ps::T, src::SourceParameters, geo::GeometricSpreadingParameters, ane::AnelasticAttenuationParameters, site::SiteParameters) where {S<:Float64,T<:Real}

Fourier acceleration spectral ordinate (m/s) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

src are the source parameters SourceParameters

geo are the geometric spreading parameters GeometricSpreadingParameters

sat are the near source saturation parameters NearSourceSaturationParameters

ane are the anelastic attenuation parameters AnelasticAttenuationParameters

site are the site parameters SiteParameters

See also: fourier_spectrum, fourier_spectrum!

fourier_spectral_ordinate(f::U, m::S, r_ps::T, src::SourceParameters, path::PathParameters, site::SiteParameters) where {S<:Real,T<:Real,U<:Float64}

Fourier acceleration spectral ordinate (m/s) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

src are the source parameters SourceParameters

path are the path parameters PathParameters

site are the site parameters SiteParameters

fourier_spectral_ordinate(f::U, m::S, r_ps::T, fas::FourierParameters) where {S<:Real,T<:Real,U<:Float64}

Fourier acceleration spectral ordinate (m/s) based upon an equivalent point source distance r_ps

f is frequency (Hz)

m is magnitude

r_ps is the equivalent point source distance including saturation effects (km)

fas are the Fourier spectral parameters FourierParameters

Individual contributions to the Fourier spectrum are also available, for example:

StochasticGroundMotionSimulation.fourier_path — Function

fourier_path(f::U, r_ps::T, m::S, geo::GeometricSpreadingParameters, sat::NearSourceSaturationParameters, ane::AnelasticAttenuationParameters) where {S<:Real,T<:Real,U<:Float64}

Path scaling of Fourier spectral model – combination of geometric spreading and anelastic attenuation.

See also: geometric_spreading, anelastic_attenuation

StochasticGroundMotionSimulation.fourier_site — Function

fourier_site(f::Float64, site::SiteParameters)

Combined site amplification and kappa filter for frequency f