Random Vibration Theory Parameters

Definition of the custom type RandomVibrationParameters to represent the model components/approaches used for the random vibration theory calculations. In particular, the type stores symbols to define the:

excitation_duration model to use,

rms_duration model to use, and

peak_factor model to use.

StochasticGroundMotionSimulation.RandomVibrationParameters — Type

RandomVibrationParameters

Struct holding parameters/methods for Random Vibration Theory.

pf_method is the method used for peak factor computation

- `:DK80` (default) is Der Kiureghian (1980), building on Vanmarcke (1975)
- `:CL56` is Cartwright Longuet-Higgins (1956)

dur_ex is the model for excitation duration

- `:BT14` (default) is the Boore & Thompson (2014) model - note that this is adpated to work with `r_ps`

dur_rms is the model for rms duration

- `:BT12` is the Boore & Thompson (2012) model
- `:BT15` (default) is the Boore & Thompson (2015) model

dur_region is the region specified for the duration model

- `:WNA` (default) is western North America
- `:ENA` is eastern North America

Note that the default specification is:

RandomVibrationParameters() = RandomVibrationParameters(:DK80, :BT14, :BT15, :WNA)

RandomVibrationParameters (generic function with 1 method)

However, an alternative constructor exists that takes a pf_method as an argument. For this constructor, the rms_duration model is linked to the peak factor method:

DK80 is paired with :BT15 as a default

CL56 is paired with :BT12 as a default

As these are currently the only two rms_duration models implemented, the constructor is specified as:

RandomVibrationParameters(pf) = RandomVibrationParameters(pf, :BT14, ((pf == :DK80) ? :BT15 : :BT12), :WNA)

RandomVibrationParameters (generic function with 1 method)

In all cases, the Boore & Thompson (2014) excitation duration model is employed as that is the only model currently implemented.

## Functionality

The main methods used to interact with RandomVibrationParameters are:

Missing docstring.

Missing docstring for spectral_moment. Check Documenter's build log for details.

Missing docstring.

Missing docstring for spectral_moments. Check Documenter's build log for details.

Missing docstring.

Missing docstring for excitation_duration. Check Documenter's build log for details.

StochasticGroundMotionSimulation.rms_duration — Function

rmsduration(m::S, rps::T, src::SourceParameters, sdof::Oscillator, rvt::RandomVibrationParameters) where {S<:Float64,T<:Real}

Returns a 3-tuple of (Drms, Dex, Dratio), using a switch on rvt.dur_rms. Default :BT12 makes use of the :BT14 model for excitation duration, Dex.

m is magnitude

r_ps is an equivalent point source distance

StochasticGroundMotionSimulation.peak_factor — Function

peak_factor(m::S, r_ps::T, fas::FourierParameters, sdof::Oscillator; pf_method::Symbol=:DK80) where {S<:Real,T<:Real}

Peak factor umax / urms with a switch of pf_method to determine the approach adopted. pf_method can currently be one of: - :CL56 for Cartright Longuet-Higgins (1956) - :DK80 for Der Kiureghian (1980), building on Vanmarcke (1975)

Defaults to :DK80.

peak_factor(m::S, r_ps::T, Dex::U, m0::V, fas::FourierParameters, sdof::Oscillator, rvt::RandomVibrationParameters) where {S<:Real,T<:Real,U<:Real,V<:Real}

Peak factor umax / urms with a switch of pf_method to determine the approach adopted. pf_method can currently be one of: - :CL56 for Cartright Longuet-Higgins (1956) - :DK80 for Der Kiureghian (1980), building on Vanmarcke (1975)

Defaults to :DK80.

Missing docstring.

Missing docstring for rvt_response_spectral_ordinate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rvt_response_spectrum. Check Documenter's build log for details.