Single Degree of Freedom Oscillator Parameters

Definition of the custom type, Oscillator to represent a single degree of freedom (SDOF) oscillator. Type simply stores the oscillator frequency and damping ratio.

Missing docstring.

Missing docstring for Oscillator. Check Documenter's build log for details.

## Functionality

StochasticGroundMotionSimulation.period — Function

period(sdof::Oscillator)

Natural period (s) of the sdof Oscillator.

StochasticGroundMotionSimulation.transfer — Function

transfer(f::T, sdof::Oscillator) where {T<:Real}

Compute the modulus of the transfer function for a SDOF system

Examples

    f = 2.0
    sdof = Oscillator(1.0, 0.05)
    Hf = transfer(f, sdof)

transfer(f::Vector{T}, sdof::Oscillator) where T<:Real

Computes the modulus of the transfer function of a SDOF for a vector of frequencies

f::Vector is the vector of frequencies

sdof::Oscillator is the oscillator instance

Examples

  f = collect(range(0.1, stop=10.0, step=0.01))
  sdof = Oscillator(1.0)
  Hf = transfer(f, sdof)

StochasticGroundMotionSimulation.squared_transfer — Function

squared_transfer(f, sdof::Oscillator)

Compute the square of the transfer function for a SDOF system, sdof, at frequency f.

Examples

f = 2.0
	# create sdof with natural frequency f_n=1.0 and damping ζ=0.05
	sdof = Oscillator( 1.0, 0.05 )
	Hf2 = squared_transfer( f, sdof )