Site Scaling

Site response is defined in terms of site amplification, or impedance effects, as well as damping – via a kappa filter.

The overall site model is therefore written as: $S(f; \bm{\theta}_S) = S_I(f) \times S_K(f)$ with $S_I(f)$ representing the impedance effects, and $S_K(f)$ being the kappa filter.

Impedance functions

Currently, three impedance functions are implemented:

:Boore2016: is the Boore (2016) impedance function for a Western US generic rock profile with $V_{S,30}=760$ m/s

:AlAtik2021_cy14: is the Al Atik & Abrahamson (2021) impedance function obtained by inverting the Chiou & Youngs (2014) GMM. The reference profile has a $V_{S,30}=760$ m/s

:Unit: simply provides a unit impedance for all frequencies, i.e., $S_I(f)=1.0$

StochasticGroundMotionSimulation.site_amplification — Function

site_amplification(f::Real, amp_model::Symbol)

Computes the site amplification (impedance) for a given frequency f. Requires the keyword argument amp_model as a String and defaults to the Boore (2016) model. Currently, any other string passed to the function will return the unit amplification

# Examples

f = 5.0
	# returns the amplification from AlAtik (2021) in both cases
	Af = site_amplification(f)
    Af = site_amplification(f; amp_model=:AlAtik2021_cy14_760)
    # returns the Boore (2016) amplification
	Af = site_amplification(f; amp_model=:Boore2016)
	# returns 1.0
	Af = site_amplification(f; amp_model=:Unit)

site_amplification(f, site::SiteParameters)

Computes the site amplification (impedance) for a given frequency f.

site_amplification(f, fas::FourierParameters)

Computes the site amplification (impedance) for a given frequency f.

Kappa filter

In addition to the impedance effects, the near surface damping is represented by a generic kappa filter:

\[S_K(f) = \exp\left( -\pi \kappa_0 f \right)\]

StochasticGroundMotionSimulation.kappa_filter — Function

kappa_filter(f, site::SiteParameters)

Kappa filter for a given frequency f